Add files via upload
n a parallel hybrid electric vehicle, both the Internal Combustion Engine (ICE) and the Electric Motor (EM) are mechanically coupled to the drivetrain and can simultaneously or independently provide traction power to the wheels. The total tractive torque is the algebraic sum of engine and motor torques, regulated by an energy management controller.